+++ src/js/core.js
@@ -164,7 +164,8 @@ const makeNewView = () => {
 		chrImportLoadVisage: false, // Whether or not to load the visage model instead (Dracthyr/Worgen)
 		chrImportChrModelID: 0, // Temporary storage for target character model ID.
 		chrImportChoices: [], // Temporary storage for character import choices.
-		characterImportMode: 'none', // Controls visibility of character import interface ('none', 'BNET')
+		chrImportWowheadURL: '', // Wowhead dressing room url
+		characterImportMode: 'none', // Controls visibility of character import interface ('none', 'BNET', 'WHEAD')
 		realmList: {}, // Contains all regions and realms once realmlist.load() has been called.
 		exportCancelled: false, // Export cancellation state.
 		isXmas: (new Date().getMonth() === 11),